move plot logic from state to render; rename files
