+++ src/components/Differential/PlotsDynamic.scss
