+++ src/components/Enrichment/EnrichmentSearch.jsx
@@ -11,7 +11,7 @@ import {
 } from 'semantic-ui-react';
 import ndjsonStream from 'can-ndjson-stream';
 import { CancelToken } from 'axios';
-import '../Shared/SearchCriteria.scss';
+import '../Shared/Search.scss';
 import { omicNavigatorService } from '../../services/omicNavigator.service';
 import EnrichmentMultisetFilters from './EnrichmentMultisetFilters';
 import { getDynamicSize } from '../Shared/helpers';
@@ -31,7 +31,7 @@ async function* streamAsyncIterable(reader) {
   }
 }
 
-class EnrichmentSearchCriteria extends Component {
+class EnrichmentSearch extends Component {
   state = {
     enrichmentStudies: [],
     enrichmentStudyHrefVisible: false,
@@ -146,7 +146,7 @@ class EnrichmentSearchCriteria extends Component {
       enrichmentAnnotation,
       pValueType,
       enrichmentTestAndDescription,
-      onSearchCriteriaChangeEnrichment,
+      onSearchChangeEnrichment,
       onSearchTransitionEnrichment,
       onGetEnrichmentsLinkouts,
     } = this.props;
@@ -257,7 +257,7 @@ class EnrichmentSearchCriteria extends Component {
             .catch(error => {
               console.error('Error during getEnrichmentsTable', error);
             });
-          onSearchCriteriaChangeEnrichment(
+          onSearchChangeEnrichment(
             {
               enrichmentStudy: enrichmentStudy,
               enrichmentModel: enrichmentModel,
@@ -280,11 +280,8 @@ class EnrichmentSearchCriteria extends Component {
   };
 
   handleStudyChange = (evt, { name, value }) => {
-    const {
-      onSearchCriteriaChangeEnrichment,
-      onSearchCriteriaResetEnrichment,
-    } = this.props;
-    onSearchCriteriaChangeEnrichment(
+    const { onSearchChangeEnrichment, onSearchResetEnrichment } = this.props;
+    onSearchChangeEnrichment(
       {
         [name]: value,
         enrichmentModel: '',
@@ -293,7 +290,7 @@ class EnrichmentSearchCriteria extends Component {
       },
       true,
     );
-    onSearchCriteriaResetEnrichment({
+    onSearchResetEnrichment({
       isValidSearchEnrichment: false,
     });
     this.setState({
@@ -351,13 +348,13 @@ class EnrichmentSearchCriteria extends Component {
   handleModelChange = (evt, { name, value }) => {
     const {
       enrichmentStudy,
-      onSearchCriteriaChangeEnrichment,
-      onSearchCriteriaResetEnrichment,
+      onSearchChangeEnrichment,
+      onSearchResetEnrichment,
       enrichmentModelsAndAnnotations,
     } = this.props;
     this.props.onHandleHasBarcodeData();
     this.props.onHandlePlotTypesEnrichment(value);
-    onSearchCriteriaChangeEnrichment(
+    onSearchChangeEnrichment(
       {
         enrichmentStudy: enrichmentStudy,
         [name]: value,
@@ -366,7 +363,7 @@ class EnrichmentSearchCriteria extends Component {
       },
       true,
     );
-    onSearchCriteriaResetEnrichment({
+    onSearchResetEnrichment({
       isValidSearchEnrichment: false,
     });
     const enrichmentModelsAndAnnotationsCopy = [
@@ -405,7 +402,7 @@ class EnrichmentSearchCriteria extends Component {
       pValueType,
       onMultisetQueriedEnrichment,
       onSearchTransitionEnrichment,
-      onSearchCriteriaChangeEnrichment,
+      onSearchChangeEnrichment,
     } = this.props;
     onSearchTransitionEnrichment(true);
     onMultisetQueriedEnrichment(false);
@@ -420,7 +417,7 @@ class EnrichmentSearchCriteria extends Component {
       reloadPlot: true,
       multisetFiltersVisibleEnrichment: false,
     });
-    onSearchCriteriaChangeEnrichment(
+    onSearchChangeEnrichment(
       {
         enrichmentStudy: enrichmentStudy,
         enrichmentModel: enrichmentModel,
@@ -693,12 +690,12 @@ class EnrichmentSearchCriteria extends Component {
       enrichmentModel,
       pValueType,
       // onSearchTransitionEnrichment,
-      onSearchCriteriaChangeEnrichment,
+      onSearchChangeEnrichment,
     } = this.props;
     this.setState({
       multisetFiltersVisibleEnrichment: false,
     });
-    onSearchCriteriaChangeEnrichment(
+    onSearchChangeEnrichment(
       {
         enrichmentStudy: enrichmentStudy,
         enrichmentModel: enrichmentModel,
@@ -1174,7 +1171,7 @@ class EnrichmentSearchCriteria extends Component {
 
     return (
       <React.Fragment>
-        <Form className="SearchCriteriaContainer">
+        <Form className="SearchContainer">
           <Popup
             trigger={
               <Form.Field
@@ -1319,4 +1316,4 @@ class EnrichmentSearchCriteria extends Component {
   }
 }
 
-export default EnrichmentSearchCriteria;
+export default EnrichmentSearch;
