+++ src/components/Shared/Search.scss
@@ -9,7 +9,7 @@
     padding-right: 10px !important;
   }
 
-  .SearchCriteriaContainer {
+  .SearchContainer {
     padding: 20px 20px 5px 20px !important;
   }
